requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
Flask-Migrate==3.1.0
python-telegram-bot==20.0a1
